[Gold IV] Title: List of Unique Numbers, Time: 12 ms, Memory: 2508 KB -BaekjoonHub

+++ 백준/Gold/13144. List of Unique Numbers/List of Unique Numbers.cc
@@ -22,7 +22,7 @@ int main()
 	
 	int N;
 	int end = 0;
-	long long res = 0;
+	long res = 0;
 	cin >> N;
 	
 	for (int i = 0; i < N; i++)
